refactor: split apollo codegen into multiple files based on service, apply formatting to unformatted files
