+++ frontend/package.json
@@ -15,6 +15,7 @@
     "@apollo/client": "^3.13.8",
     "@apollo/client-integration-nextjs": "^0.12.2",
     "@apollo/experimental-nextjs-app-support": "^0.12.2",
+    "@graphql-codegen/near-operation-file-preset": "^3.1.0",
     "@next/eslint-plugin-next": "^15.3.3",
     "@radix-ui/react-dialog": "^1.1.14",
     "@radix-ui/react-label": "^2.1.7",
