+++ frontend/src/app/dashboard/page.tsx
@@ -1,10 +1,8 @@
 import MyGrid from '../../components/dashboard/TestGrid';
-import DashboardWelcomeMessage from '../../components/dashboard/WelcomeMessage';
 
 export default function Dashboard() {
     return (
         <div>
-            <DashboardWelcomeMessage />
             <MyGrid />
         </div>
     );
