+++ frontend/src/components/dashboard/CoordinateGrid.tsx
@@ -1,4 +1,4 @@
-import { Widget } from '@/generated/graphql';
+import { Widget } from '@/generated/types';
 
 interface CoordinateGridProps {
     showCoordinates: boolean;
