+++ frontend/src/components/dashboard/WidgetActions.tsx
@@ -9,7 +9,8 @@ import {
     DialogHeader,
     DialogTitle,
 } from '@/components/ui/dialog';
-import { Widget, useDeleteWidgetMutation } from '@/generated/graphql';
+import { Widget } from '@/generated/types';
+import { useDeleteWidgetMutation } from '@/generated/Widgets.generated';
 import { getWidgetIconStyles } from '@/lib/utils/widgetStyles';
 import { Edit2, Palette, Trash2 } from 'lucide-react';
 import { useState } from 'react';
