+++ frontend/src/components/dashboard/widgets/NotesWidget.tsx
@@ -5,7 +5,6 @@ import { Badge } from '@/components/ui/badge';
 import { Button } from '@/components/ui/button';
 import { Input } from '@/components/ui/input';
 import { MarkdownRenderer } from '@/components/ui/markdown-renderer';
-import { NotesConfig, Widget } from '@/generated/graphql';
 import {
     getWidgetClasses,
     getWidgetIconStyles,
@@ -28,6 +27,7 @@ import { WidgetActions } from '../WidgetActions';
 import { useNoteLayout } from '../hooks/useNoteLayout';
 import { NoteCard } from './NoteCard';
 import { NoteModal } from './NoteModal';
+import { Widget, NotesConfig } from '@/generated/types';
 
 interface NotesWidgetProps {
     widget: Widget;
