+++ frontend/src/lib/graphql/Auth.graphql
@@ -1,3 +1,10 @@
+query Me {
+    me {
+        email
+        name
+    }
+}
+
 mutation Refresh {
     refresh {
         message
