--- frontend/src/lib/graphql/Me.graphql
@@ -1,6 +1,0 @@
-query Me {
-    me {
-        email
-        name
-    }
-}
