+++ frontend/src/lib/utils/widget-styling/icon-styles.ts
@@ -1,4 +1,4 @@
-import { Widget } from '@/generated/graphql';
+import { Widget } from '@/generated/types';
 import { WidgetIconStyles } from './types';
 
 /**
