+++ frontend/src/lib/utils/widget-styling/widget-styles.ts
@@ -1,4 +1,4 @@
-import { Widget } from '@/generated/graphql';
+import { Widget } from '@/generated/types';
 import { applyBackgroundStyles } from './background-strategy';
 
 /**
